feat: settings system, codec/executor expansion, histogram ms precision, health panel
Replace TOML config loading with file-based settings system (flowcus.settings)
that supports runtime updates and a settings API. Add SettingsPanel and
HealthPanel frontend components. Remove dev-only profiling module and
storage metrics in favor of production-ready observability.

Major changes:
- New settings system: core settings loader, server settings API, restart support
- Storage codec expansion: extended transform codecs with auto-selection
- Executor overhaul: expanded query execution with histogram improvements
- Histogram timestamps changed from seconds to milliseconds (bucket_ms)
- Tuned storage defaults: 16MB flush, 5s interval, 8192 channel, 4 merge workers
- Part format improvements and merge coordinator updates
- Frontend: HealthPanel, SettingsPanel, updated histogram/gantt components
- Removed: profiling.rs, metrics.rs, auto-commit hook

+++ crates/flowcus-storage/src/table.rs
@@ -33,8 +33,8 @@ pub struct PartEntry {
     /// Part format version (0 = legacy 4-segment dir name, 1 = current 5-segment).
     pub format_version: u32,
     pub generation: u32,
-    pub time_min: u32,
-    pub time_max: u32,
+    pub time_min: u64,
+    pub time_max: u64,
     pub seq: u32,
 }
 
@@ -65,8 +65,8 @@ impl Table {
     /// Returns parts sorted by (generation, time_min, seq).
     pub fn list_parts(
         &self,
-        time_min: Option<u32>,
-        time_max: Option<u32>,
+        time_min: Option<u64>,
+        time_max: Option<u64>,
     ) -> std::io::Result<Vec<PartEntry>> {
         let mut parts = Vec::new();
         walk_parts_recursive(&self.base_dir, 0, time_min, time_max, &mut parts)?;
@@ -88,8 +88,8 @@ impl Table {
 fn walk_parts_recursive(
     dir: &Path,
     depth: u8,
-    time_min: Option<u32>,
-    time_max: Option<u32>,
+    time_min: Option<u64>,
+    time_max: Option<u64>,
     out: &mut Vec<PartEntry>,
 ) -> std::io::Result<()> {
     if !dir.exists() {
